[Refactor]: 공용 kakaoMap을 활용한 별도의 RegisterMap 구현에 따른 import 경로 변경 및 파일변경
