+++ src/app/(no-nav)/register/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { RegisterHeader } from "@/components/header";
-import RegisterMap from "@/components/register/mapSection/RegisterMap";
+import RegisterMap from "@/features/map/register/ui/RegisterMap";
 import RegisterForm from "@/components/register/registerForm/RegisterForm";
-import { useEnsureMeasurement } from "@/hook/useEnsureMeasurement";
+import { useEnsureMeasurement } from "@/features/measurement/model/hooks/useEnsureMeasurement";
 
 export default function Register() {
     useEnsureMeasurement();
